backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
scikit-learn
keybert
rake-nltk
sentence-transformers
